lca_update & FLIC
We found out that the LCA reported Heywood cases. We redid the analysis with in the checkHess function TRUE so that the solutions were checked for applicable variance matrices for the parameters. The three class solution was still the best solution and the probability plots show no big differences.

We tried to finish the distal outcomes analysis, but encountered new problems to overcome. The negative weights received from the inverse D matrix made it impossible to apply these weights to a Frith Regression analysis. We tried the ad hoc fixes of fixing the negative weights to zero or 0.01. Both results are reported. The first incorporates only the original number of participants with their modal class predictions, with different weights per class. The second approach uses all of the expanded data set and probably therefore also underestimates the SEs of the parameters.