IRLS for lr tests
We found out that the automatically reported significance tests for the flic regressions are faulty and needed to be done manually. We encountered a problem for model 5 (it was impossible to estimate a null model to test against), which was fixed by employing another optimization algorithm (IRLS instead of Newton-Raphson). The Flic models were refit using the IRLS algorithm.