Extra Analysis
It was decided that an extra analysis is needed to explore the class difference for specifically people who had a concept of "young people" that was included an age below 15 years old.

The results were added to the BCH_results and BCH_results_parameters tables. The exosting tables were overwritten.